Genomic Tables

Generates tables of gene names, log2 values, p-values. One can filter for genes that were up or down regulated, or the default is neither. Function will prompt for the number of rows you wish to view.

genomic_tables(
  data,
  U.D.diffexp = c("neither", "up", "down"),
  gene.n,
  log2,
  pval,
  sig,
  num.rows = 10
)

Arguments

data
name of data set

U.D.diffexp
pick either "up" or "down" to filter for genes in those categories

gene.n
column of data set with gene names

log2
column of data set with log2 values

pval
column of data set with p-values

sig
column name in data set that states changes were significant

num.rows
number of rows to display in the table

Value

returns a huxtable

Examples

genomic_tables(data = gendata, U.D.diffexp = "up", gene.n = "gene",
       log2 = "log2.fold_change.", pval = "p_value", sig = "significant")

#>                     gene    log2.fold_change.       p_value  
#>                   ───────────────────────────────────────────
#>                     4759                  Inf       0.001    
#>                     6760                 2.98   5e-05        
#>                     1697                 2.97   5e-05        
#>                     6518                 2.94   5e-05        
#>                     5030                 2.56       0.00065  
#>                     7271                 2.42   5e-05        
#>                     2776                 2.39   5e-05        
#>                     1629                  2.3   5e-05        
#>                     7441                 2.16   5e-05        
#>                     7272                 2.13   5e-05        
#> 
#> Column names: gene, log2.fold_change., p_value

genomic_tables(data = gendata, U.D.diffexp = "down", gene.n = "gene",
       log2 = "log2.fold_change.", pval = "p_value", sig = "significant")

#>                     gene    log2.fold_change.       p_value  
#>                   ───────────────────────────────────────────
#>                     16938               -3.99   5e-05        
#>                     6121                 -3.9       0.0003   
#>                     5945                   -3       0.00215  
#>                     5656                -2.39   5e-05        
#>                     2864                -2.32       0.0004   
#>                     8020                -2.11   5e-05        
#>                     7619                 -1.8   5e-05        
#>                     5504                -1.75   5e-05        
#>                     5655                -1.75   5e-05        
#>                     8215                 -1.7       0.0017   
#> 
#> Column names: gene, log2.fold_change., p_value

genomic_tables(data = gendata, U.D.diffexp = "neither", gene.n = "gene",
       log2 = "log2.fold_change.", pval = "p_value", sig = "significant")

#>              gene    log2.fold_change       p_value   diffexpressed  
#>                                     .                                
#>            ──────────────────────────────────────────────────────────
#>              16938              -3.99   5e-05         down           
#>              185                0.706       0.0001    up             
#>              16645              0.556       0.0012    up             
#>              207               -0.698       0.00055   down           
#>              296                 1.39   5e-05         up             
#>              452                0.871       0.001     up             
#>              485               -0.863       0.0006    down           
#>              597                -1.26   5e-05         down           
#>              599                -1.23       0.0004    down           
#>              635                0.591       0.00105   up             
#> 
#> Column names: gene, log2.fold_change., p_value, diffexpressed

genomic_tables(data = gendata, gene.n = "gene", log2 = "log2.fold_change.",
        pval = "p_value", sig = "significant")

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#>              gene    log2.fold_change       p_value   diffexpressed  
#>                                     .                                
#>            ──────────────────────────────────────────────────────────
#>              16938              -3.99   5e-05         down           
#>              185                0.706       0.0001    up             
#>              16645              0.556       0.0012    up             
#>              207               -0.698       0.00055   down           
#>              296                 1.39   5e-05         up             
#>              452                0.871       0.001     up             
#>              485               -0.863       0.0006    down           
#>              597                -1.26   5e-05         down           
#>              599                -1.23       0.0004    down           
#>              635                0.591       0.00105   up             
#> 
#> Column names: gene, log2.fold_change., p_value, diffexpressed